+++ cloudos_cli/__main__.py
@@ -22,6 +22,8 @@ from cloudos_cli.utils.array_job import generate_datasets_for_project
 from cloudos_cli.utils.details import get_path
 from cloudos_cli.link import Link
 from cloudos_cli.cost.cost import CostViewer
+from cloudos_cli.logging.logger import setup_logging, update_command_context_from_click
+import logging
 
 
 # GLOBAL VARS
@@ -41,19 +43,22 @@ INIT_PROFILE = 'initialisingProfile'
 def handle_exception(show_traceback=False):
     """Custom exception handler for CloudOS CLI"""
     def exception_handler(exc_type, exc_value, exc_traceback):
+        # Initialise logger
+        debug_mode = '--debug' in sys.argv
+        setup_logging(debug_mode)
+        logger = logging.getLogger("CloudOS")
         # Handle keyboard interrupt gracefully
         if issubclass(exc_type, KeyboardInterrupt):
             sys.exit(1)
-        
         if show_traceback:
             # Show full traceback for debugging
+            logger.error(exc_value, exc_info=exc_value)
             traceback.print_exception(exc_type, exc_value, exc_traceback)
         else:
             # Show only the error message
+            logger.error(exc_value)
             click.echo(click.style(f"Error: {exc_value}", fg='red'), err=True)
-        
         sys.exit(1)
-    
     return exception_handler
 
 
@@ -63,15 +68,14 @@ def handle_exception(show_traceback=False):
 @click.pass_context
 def run_cloudos_cli(ctx, debug):
     """CloudOS python package: a package for interacting with CloudOS."""
+    update_command_context_from_click(ctx)
     ctx.ensure_object(dict)
     ctx.obj['debug'] = debug
-    
     # Set up custom exception handling based on debug flag
     if not debug:
         sys.excepthook = handle_exception(show_traceback=False)
     else:
         sys.excepthook = handle_exception(show_traceback=True)
-    
     if ctx.invoked_subcommand not in ['datasets']:
         print(run_cloudos_cli.__doc__ + '\n')
         print('Version: ' + __version__ + '\n')
@@ -263,6 +267,7 @@ def images():
 @click.pass_context
 def datasets(ctx):
     """CloudOS datasets functionality."""
+    update_command_context_from_click(ctx)
     if ctx.args and ctx.args[0] != 'ls':
         print(datasets.__doc__ + '\n')
 
@@ -276,6 +281,7 @@ def datasets(ctx):
 def configure(ctx, profile, make_default):
     """CloudOS configuration."""
     print(configure.__doc__ + '\n')
+    update_command_context_from_click(ctx)
     profile = profile or ctx.obj['profile']
     config_manager = ConfigurationProfile()
 
@@ -286,6 +292,7 @@ def configure(ctx, profile, make_default):
         config_manager.create_profile_from_input(profile_name=profile)
     if make_default:
         config_manager.make_default_profile(profile_name=profile)
+
 
 @job.command('run')
 @click.option('-k',
@@ -477,6 +484,7 @@ def run(ctx,
         ssl_cert,
         profile):
     """Submit a job to CloudOS."""
+    update_command_context_from_click(ctx)
     profile = profile or ctx.default_map['job']['run']['profile']
     # Create a dictionary with required and non-required params
     required_dict = {
@@ -484,6 +492,7 @@ def run(ctx,
         'workspace_id': True,
         'workflow_name': True,
         'project_name': True,
+        'session_id': False,
         'procurement_id': False
     }
     # determine if the user provided all required parameters
@@ -768,6 +777,7 @@ def job_status(ctx,
                ssl_cert,
                profile):
     """Check job status in CloudOS."""
+    update_command_context_from_click(ctx)
     profile = profile or ctx.default_map['job']['status']['profile']
     # Create a dictionary with required and non-required params
     required_dict = {
@@ -775,6 +785,7 @@ def job_status(ctx,
         'workspace_id': False,
         'workflow_name': False,
         'project_name': False,
+        'session_id': False,
         'procurement_id': False
     }
     # determine if the user provided all required parameters
@@ -802,12 +813,19 @@ def job_status(ctx,
         print('\tThe following Cloudos object was created:')
         print('\t' + str(cl) + '\n')
         print(f'\tSearching for job id: {job_id}')
-    j_status = cl.get_job_status(job_id, verify_ssl)
-    j_status_h = json.loads(j_status.content)["status"]
-    print(f'\tYour current job status is: {j_status_h}\n')
-    j_url = f'{cloudos_url}/app/advanced-analytics/analyses/{job_id}'
-    print(f'\tTo further check your job status you can either go to {j_url} ' +
-          'or repeat the command you just used.')
+    try:
+        j_status = cl.get_job_status(job_id, verify_ssl)
+        j_status_h = json.loads(j_status.content)["status"]
+        print(f'\tYour current job status is: {j_status_h}\n')
+        j_url = f'{cloudos_url}/app/advanced-analytics/analyses/{job_id}'
+        print(f'\tTo further check your job status you can either go to {j_url} ' +
+              'or repeat the command you just used.')
+    except BadRequestException as e:
+        click.echo(f"[ERROR] Job '{job_id}' not found or not accessible: {str(e)}", err=True)
+        sys.exit(1)
+    except Exception as e:
+        click.echo(f"[ERROR] Failed to retrieve status for job '{job_id}': {str(e)}", err=True)
+        sys.exit(1)
 
 
 @job.command('workdir')
@@ -838,15 +856,16 @@ def job_status(ctx,
 @click.option('--profile', help='Profile to use from the config file', default=None)
 @click.pass_context
 def job_workdir(ctx,
-             apikey,
-             cloudos_url,
-             workspace_id,
-             job_id,
-             verbose,
-             disable_ssl_verification,
-             ssl_cert,
-             profile):
+                apikey,
+                cloudos_url,
+                workspace_id,
+                job_id,
+                verbose,
+                disable_ssl_verification,
+                ssl_cert,
+                profile):
     """Get the path to the working directory of a specified job."""
+    update_command_context_from_click(ctx)
     profile = profile or ctx.default_map['job']['workdir']['profile']
     # Create a dictionary with required and non-required params
     required_dict = {
@@ -854,6 +873,7 @@ def job_workdir(ctx,
         'workspace_id': True,
         'workflow_name': False,
         'project_name': False,
+        'session_id': False,
         'procurement_id': False
     }
     # determine if the user provided all required parameters
@@ -883,8 +903,15 @@ def job_workdir(ctx,
         print('\tThe following Cloudos object was created:')
         print('\t' + str(cl) + '\n')
         print(f'\tSearching for job id: {job_id}')
-    workdir = cl.get_job_workdir(job_id, workspace_id, verify_ssl)
-    print(f"Working directory for job {job_id}: {workdir}")
+    try:
+        workdir = cl.get_job_workdir(job_id, workspace_id, verify_ssl)
+        print(f"Working directory for job {job_id}: {workdir}")
+    except BadRequestException as e:
+        click.echo(f"[ERROR] Job '{job_id}' not found or not accessible: {str(e)}", err=True)
+        sys.exit(1)
+    except Exception as e:
+        click.echo(f"[ERROR] Failed to retrieve working directory for job '{job_id}': {str(e)}", err=True)
+        sys.exit(1)
 
 
 @job.command('logs')
@@ -924,6 +951,7 @@ def job_logs(ctx,
              ssl_cert,
              profile):
     """Get the path to the logs of a specified job."""
+    update_command_context_from_click(ctx)
     profile = profile or ctx.default_map['job']['logs']['profile']
     # Create a dictionary with required and non-required params
     required_dict = {
@@ -931,6 +959,7 @@ def job_logs(ctx,
         'workspace_id': True,
         'workflow_name': False,
         'project_name': False,
+        'session_id': False,
         'procurement_id': False
     }
     # determine if the user provided all required parameters
@@ -960,9 +989,16 @@ def job_logs(ctx,
         print('\tThe following Cloudos object was created:')
         print('\t' + str(cl) + '\n')
         print(f'\tSearching for job id: {job_id}')
-    logs = cl.get_job_logs(job_id, workspace_id, verify_ssl)
-    for name, path in logs.items():
-        print(f"{name}: {path}\n")
+    try:
+        logs = cl.get_job_logs(job_id, workspace_id, verify_ssl)
+        for name, path in logs.items():
+            print(f"{name}: {path}\n")
+    except BadRequestException as e:
+        click.echo(f"[ERROR] Job '{job_id}' not found or not accessible: {str(e)}", err=True)
+        sys.exit(1)
+    except Exception as e:
+        click.echo(f"[ERROR] Failed to retrieve logs for job '{job_id}': {str(e)}", err=True)
+        sys.exit(1)
 
 
 @job.command('results')
@@ -1002,6 +1038,7 @@ def job_results(ctx,
                 ssl_cert,
                 profile):
     """Get the path to the results of a specified job."""
+    update_command_context_from_click(ctx)
     profile = profile or ctx.default_map['job']['results']['profile']
     # Create a dictionary with required and non-required params
     required_dict = {
@@ -1009,6 +1046,7 @@ def job_results(ctx,
         'workspace_id': True,
         'workflow_name': False,
         'project_name': False,
+        'session_id': False,
         'procurement_id': False
     }
     # determine if the user provided all required parameters
@@ -1038,9 +1076,16 @@ def job_results(ctx,
         print('\tThe following Cloudos object was created:')
         print('\t' + str(cl) + '\n')
         print(f'\tSearching for job id: {job_id}')
-    logs = cl.get_job_results(job_id, workspace_id, verify_ssl)
-    for name, path in logs.items():
-        print(f"{name}: {path}\n")
+    try:
+        results = cl.get_job_results(job_id, workspace_id, verify_ssl)
+        for name, path in results.items():
+            print(f"{name}: {path}\n")
+    except BadRequestException as e:
+        click.echo(f"[ERROR] Job '{job_id}' not found or not accessible: {str(e)}", err=True)
+        sys.exit(1)
+    except Exception as e:
+        click.echo(f"[ERROR] Failed to retrieve results for job '{job_id}': {str(e)}", err=True)
+        sys.exit(1)
 
 
 @job.command('details')
@@ -1093,6 +1138,7 @@ def job_details(ctx,
                 ssl_cert,
                 profile):
     """Retrieve job details in CloudOS."""
+    update_command_context_from_click(ctx)
     profile = profile or ctx.default_map['job']['details']['profile']
     # Create a dictionary with required and non-required params
     required_dict = {
@@ -1100,6 +1146,7 @@ def job_details(ctx,
         'workspace_id': False,
         'workflow_name': False,
         'project_name': False,
+        'session_id': False,
         'procurement_id': False
     }
     # determine if the user provided all required parameters
@@ -1136,6 +1183,12 @@ def job_details(ctx,
         if '403' in str(e) or 'Forbidden' in str(e):
             print("[Error] API can only show job details of your own jobs, cannot see other user's job details.")
             sys.exit(1)
+        else:
+            click.echo(f"[ERROR] Job '{job_id}' not found or not accessible: {str(e)}", err=True)
+            sys.exit(1)
+    except Exception as e:
+        click.echo(f"[ERROR] Failed to retrieve details for job '{job_id}': {str(e)}", err=True)
+        sys.exit(1)
     j_details_h = json.loads(j_details.content)
 
     # Determine the execution platform based on jobType
@@ -1330,7 +1383,7 @@ def job_details(ctx,
 @click.option('--filter-queue',
               help='Filter jobs by queue name. Only applies to jobs running in batch environment. Non-batch jobs are preserved in results.')
 @click.option('--filter-owner',
-             help='Filter jobs by owner username.')
+              help='Filter jobs by owner username.')
 @click.option('--verbose',
               help='Whether to print information messages or not.',
               is_flag=True)
@@ -1366,6 +1419,7 @@ def list_jobs(ctx,
               ssl_cert,
               profile):
     """Collect workspace jobs from a CloudOS workspace in CSV or JSON format."""
+    update_command_context_from_click(ctx)
     profile = profile or ctx.default_map['job']['list']['profile']
     # Create a dictionary with required and non-required params
     required_dict = {
@@ -1373,6 +1427,7 @@ def list_jobs(ctx,
         'workspace_id': True,
         'workflow_name': False,
         'project_name': False,
+        'session_id': False,
         'procurement_id': False
     }
     # determine if the user provided all required parameters
@@ -1437,7 +1492,6 @@ def list_jobs(ctx,
             filter_owner,
             filter_queue
         ])
-        
         if filters_used:
             print('\t[Message] A total of 0 jobs collected.')
         elif ctx.get_parameter_source('page') == click.core.ParameterSource.DEFAULT:
@@ -1498,6 +1552,7 @@ def abort_jobs(ctx,
                ssl_cert,
                profile):
     """Abort all specified jobs from a CloudOS workspace."""
+    update_command_context_from_click(ctx)
     profile = profile or ctx.default_map['job']['abort']['profile']
     # Create a dictionary with required and non-required params
     required_dict = {
@@ -1505,6 +1560,7 @@ def abort_jobs(ctx,
         'workspace_id': True,
         'workflow_name': False,
         'project_name': False,
+        'session_id': False,
         'procurement_id': False
     }
     # determine if the user provided all required parameters
@@ -1680,7 +1736,7 @@ def job_cost(ctx,
               is_flag=True)
 @click.option('--job-queue',
               help=('Name of the job queue to use with a batch job. ' +
-                   'In Azure workspaces, this option is ignored.'))
+                    'In Azure workspaces, this option is ignored.'))
 @click.option('--instance-type',
               help=('The type of compute instance to use as master node. ' +
                     'Default=c5.xlarge(aws)|Standard_D4as_v4(azure).'))
@@ -1710,27 +1766,27 @@ def job_cost(ctx,
               default=None)
 @click.pass_context
 def clone_resume(ctx,
-              apikey,
-              cloudos_url,
-              workspace_id,
-              project_name,
-              parameter,
-              nextflow_profile,
-              nextflow_version,
-              git_branch,
-              job_name,
-              do_not_save_logs,
-              job_queue,
-              instance_type,
-              cost_limit,
-              job_id,
-              accelerate_file_staging,
-              resumable,
-              verbose,
-              disable_ssl_verification,
-              ssl_cert,
-              profile):
-    """Clone/resume an existing job with optional parameter overrides."""
+                 apikey,
+                 cloudos_url,
+                 workspace_id,
+                 project_name,
+                 parameter,
+                 nextflow_profile,
+                 nextflow_version,
+                 git_branch,
+                 job_name,
+                 do_not_save_logs,
+                 job_queue,
+                 instance_type,
+                 cost_limit,
+                 job_id,
+                 accelerate_file_staging,
+                 resumable,
+                 verbose,
+                 disable_ssl_verification,
+                 ssl_cert,
+                 profile):
+    update_command_context_from_click(ctx)
     if ctx.info_name == "clone":
         mode, action = "clone", "cloning"
     elif ctx.info_name == "resume":
@@ -1744,6 +1800,7 @@ def clone_resume(ctx,
         'workspace_id': True,
         'workflow_name': False,
         'project_name': False,
+        'session_id': False,
         'procurement_id': False
     }
 
@@ -1774,7 +1831,7 @@ def clone_resume(ctx,
 
     # Create Job object (set dummy values for project_name and workflow_name, since they come from the cloned job)
     job_obj = jb.Job(cloudos_url, apikey, None, workspace_id, None, None, workflow_id=1234, project_id="None",
-               mainfile=None, importsfile=None,verify=verify_ssl)
+                     mainfile=None, importsfile=None, verify=verify_ssl)
 
     if verbose:
         print('\tThe following Job object was created:')
@@ -1807,13 +1864,13 @@ def clone_resume(ctx,
         print(f"Job successfully {mode}d. New job ID: {cloned_resumed_job_id}")
 
     except BadRequestException as e:
-        if verbose:
-            print(f'\tError details: {e}')
-        raise ValueError(f"Failed to {mode} job: {e}")
+        click.echo(f"[ERROR] Failed to {mode} job.  Job '{job_id}' not found or not accessible: {str(e)}", err=True)
+        sys.exit(1)
     except Exception as e:
-        if verbose:
-            print(f'\tError details: {e}')
-        raise ValueError(f"An error occurred while {action} the job: {e}")
+        click.echo(f"[ERROR] Failed to {mode} job. Failed to {action} job '{job_id}': {str(e)}", err=True)
+        sys.exit(1)
+
+
 # Register the same function under two names
 job.add_command(clone_resume, "clone")
 job.add_command(clone_resume, "resume")
@@ -1869,6 +1926,7 @@ def list_workflows(ctx,
                    ssl_cert,
                    profile):
     """Collect all workflows from a CloudOS workspace in CSV format."""
+    update_command_context_from_click(ctx)
     profile = profile or ctx.default_map['workflow']['list']['profile']
     # Create a dictionary with required and non-required params
     required_dict = {
@@ -1876,6 +1934,7 @@ def list_workflows(ctx,
         'workspace_id': True,
         'workflow_name': False,
         'project_name': False,
+        'session_id': False,
         'procurement_id': False
     }
     # determine if the user provided all required parameters
@@ -1967,6 +2026,7 @@ def import_wf(ctx,
     """
     Import workflows from supported repository providers.
     """
+    update_command_context_from_click(ctx)
     profile = profile or ctx.default_map['workflow']['import']['profile']
     # Create a dictionary with required and non-required params
     required_dict = {
@@ -1974,6 +2034,7 @@ def import_wf(ctx,
         'workspace_id': True,
         'workflow_name': True,
         'project_name': False,
+        'session_id': False,
         'procurement_id': False
     }
     # determine if the user provided all required parameters
@@ -2064,6 +2125,7 @@ def list_projects(ctx,
                   ssl_cert,
                   profile):
     """Collect all projects from a CloudOS workspace in CSV format."""
+    update_command_context_from_click(ctx)
     profile = profile or ctx.default_map['project']['list']['profile']
     # Create a dictionary with required and non-required params
     required_dict = {
@@ -2071,6 +2133,7 @@ def list_projects(ctx,
         'workspace_id': True,
         'workflow_name': False,
         'project_name': False,
+        'session_id': False,
         'procurement_id': False
     }
     # determine if the user provided all required parameters
@@ -2169,6 +2232,7 @@ def create_project(ctx,
                    ssl_cert,
                    profile):
     """Create a new project in CloudOS."""
+    update_command_context_from_click(ctx)
     profile = profile or ctx.default_map['project']['create']['profile']
     # Create a dictionary with required and non-required params
     required_dict = {
@@ -2176,6 +2240,7 @@ def create_project(ctx,
         'workspace_id': True,
         'workflow_name': False,
         'project_name': False,
+        'session_id': False,
         'procurement_id': False
     }
     # determine if the user provided all required parameters
@@ -2257,6 +2322,7 @@ def cromwell_status(ctx,
                     ssl_cert,
                     profile):
     """Check Cromwell server status in CloudOS."""
+    update_command_context_from_click(ctx)
     profile = profile or ctx.default_map['cromwell']['status']['profile']
     # Create a dictionary with required and non-required params
     required_dict = {
@@ -2264,6 +2330,7 @@ def cromwell_status(ctx,
         'workspace_id': True,
         'workflow_name': False,
         'project_name': False,
+        'session_id': False,
         'procurement_id': False
     }
     # determine if the user provided all required parameters
@@ -2342,6 +2409,7 @@ def cromwell_restart(ctx,
                      ssl_cert,
                      profile):
     """Restart Cromwell server in CloudOS."""
+    update_command_context_from_click(ctx)
     profile = profile or ctx.default_map['cromwell']['status']['profile']
     # Create a dictionary with required and non-required params
     required_dict = {
@@ -2349,6 +2417,7 @@ def cromwell_restart(ctx,
         'workspace_id': True,
         'workflow_name': False,
         'project_name': False,
+        'session_id': False,
         'procurement_id': False
     }
     # determine if the user provided all required parameters
@@ -2444,6 +2513,7 @@ def cromwell_stop(ctx,
                   ssl_cert,
                   profile):
     """Stop Cromwell server in CloudOS."""
+    update_command_context_from_click(ctx)
     profile = profile or ctx.default_map['cromwell']['status']['profile']
     # Create a dictionary with required and non-required params
     required_dict = {
@@ -2451,6 +2521,7 @@ def cromwell_stop(ctx,
         'workspace_id': True,
         'workflow_name': False,
         'project_name': False,
+        'session_id': False,
         'procurement_id': False
     }
     # determine if the user provided all required parameters
@@ -2534,6 +2605,7 @@ def list_queues(ctx,
                 ssl_cert,
                 profile):
     """Collect all available job queues from a CloudOS workspace."""
+    update_command_context_from_click(ctx)
     profile = profile or ctx.default_map['queue']['list']['profile']
     # Create a dictionary with required and non-required params
     required_dict = {
@@ -2541,6 +2613,7 @@ def list_queues(ctx,
         'workspace_id': True,
         'workflow_name': False,
         'project_name': False,
+        'session_id': False,
         'procurement_id': False
     }
     # determine if the user provided all required parameters
@@ -2593,6 +2666,7 @@ def list_profiles():
               required=True)
 @click.pass_context
 def remove_profile(ctx, profile):
+    update_command_context_from_click(ctx)
     profile = profile or ctx.obj['profile']
     config_manager = ConfigurationProfile()
     config_manager.remove_profile(profile)
@@ -2724,6 +2798,7 @@ def run_bash_job(ctx,
                  ssl_cert,
                  profile):
     """Run a bash job in CloudOS."""
+    update_command_context_from_click(ctx)
     profile = profile or ctx.default_map['bash']['job']['profile']
     # Create a dictionary with required and non-required params
     required_dict = {
@@ -2731,6 +2806,7 @@ def run_bash_job(ctx,
         'workspace_id': True,
         'workflow_name': True,
         'project_name': True,
+        'session_id': False,
         'procurement_id': False
     }
 
@@ -3011,6 +3087,7 @@ def run_bash_array_job(ctx,
                        custom_script_path,
                        custom_script_project):
     """Run a bash array job in CloudOS."""
+    update_command_context_from_click(ctx)
     profile = profile or ctx.default_map['bash']['array-job']['profile']
 
     # Create a dictionary with required and non-required params
@@ -3019,6 +3096,7 @@ def run_bash_array_job(ctx,
         'workspace_id': True,
         'workflow_name': True,
         'project_name': True,
+        'session_id': False,
         'procurement_id': False
     }
 
@@ -3063,11 +3141,11 @@ def run_bash_array_job(ctx,
 
     # setup separators for API and array file (the're different)
     separators = {
-        ",": { "api": ",", "file": "," },
-        ";": { "api": "%3B", "file": ";" },
-        "space": { "api": "+", "file": " " },
-        "tab": { "api": "tab", "file": "tab" },
-        "|": { "api": "%7C", "file": "|" }
+        ",": {"api": ",", "file": ","},
+        ";": {"api": "%3B", "file": ";"},
+        "space": {"api": "+", "file": " "},
+        "tab": {"api": "tab", "file": "tab"},
+        "|": {"api": "%7C", "file": "|"}
     }
 
     # setup important options for the job
@@ -3241,7 +3319,7 @@ def list_files(ctx,
                path,
                details):
     """List contents of a path within a CloudOS workspace dataset."""
-
+    update_command_context_from_click(ctx)
     profile = profile or ctx.default_map['datasets']['list'].get('profile')
     config_manager = ConfigurationProfile()
 
@@ -3250,6 +3328,7 @@ def list_files(ctx,
         'workspace_id': True,
         'workflow_name': False,
         'project_name': False,
+        'session_id': False,
         'procurement_id': False
     }
 
@@ -3379,9 +3458,8 @@ def move_files(ctx, source_path, destination_path, apikey, cloudos_url, workspac
     DESTINATION_PATH [path]: the full path to the destination folder. It must be a 'Data' folder path.
      E.g.: 'Data/folderB'
     """
-
+    update_command_context_from_click(ctx)
     profile = profile or ctx.default_map['datasets']['move'].get('profile')
-    
     # Validate destination constraint
     if not destination_path.strip("/").startswith("Data/") and destination_path.strip("/") != "Data":
         click.echo("[ERROR] Destination path must begin with 'Data/' or be 'Data'.", err=True)
@@ -3397,6 +3475,7 @@ def move_files(ctx, source_path, destination_path, apikey, cloudos_url, workspac
         'workspace_id': True,
         'workflow_name': False,
         'project_name': True,
+        'session_id': False,
         'procurement_id': False
     }
 
@@ -3486,9 +3565,10 @@ def move_files(ctx, source_path, destination_path, apikey, cloudos_url, workspac
         # Normalize kind: top-level datasets are kind=Dataset, all other folders are kind=Folder
         if folder_type in ("VirtualFolder", "Folder"):
             target_kind = "Folder"
-        elif folder_type=="S3Folder":
-            click.echo(f"[ERROR] Unable to move item '{source_item_name}' to '{destination_path}'. The destination is an S3 folder, and only virtual folders can be selected as valid move destinations.",
-                   err=True)
+        elif folder_type == "S3Folder":
+            click.echo(f"[ERROR] Unable to move item '{source_item_name}' to '{destination_path}'. " +
+                       "The destination is an S3 folder, and only virtual folders can be selected as valid move destinations.",
+                       err=True)
             sys.exit(1)
         elif isinstance(folder_type, bool) and folder_type:  # legacy dataset structure
             target_kind = "Dataset"
@@ -3530,9 +3610,16 @@ def move_files(ctx, source_path, destination_path, apikey, cloudos_url, workspac
 @click.option('--ssl-cert', help='Path to your SSL certificate file.')
 @click.option('--profile', default=None, help='Profile to use from the config file.')
 @click.pass_context
-def renaming_item(ctx, source_path, new_name, apikey, cloudos_url,
-                    workspace_id, project_name,
-                    disable_ssl_verification, ssl_cert, profile):
+def renaming_item(ctx,
+                  source_path,
+                  new_name,
+                  apikey,
+                  cloudos_url,
+                  workspace_id,
+                  project_name,
+                  disable_ssl_verification,
+                  ssl_cert,
+                  profile):
     """
     Rename a file or folder in a CloudOS project.
 
@@ -3540,6 +3627,7 @@ def renaming_item(ctx, source_path, new_name, apikey, cloudos_url,
      E.g.: 'Data/folderA/old_name.txt'\n
     NEW_NAME [name]: the new name to assign to the file or folder. E.g.: 'new_name.txt'
     """
+    update_command_context_from_click(ctx)
     if not source_path.strip("/").startswith("Data/"):
         click.echo("[ERROR] SOURCE_PATH must start with 'Data/', pointing to a file/folder in that dataset.", err=True)
         sys.exit(1)
@@ -3550,6 +3638,7 @@ def renaming_item(ctx, source_path, new_name, apikey, cloudos_url,
         'workspace_id': True,
         'workflow_name': False,
         'project_name': True,
+        'session_id': False,
         'procurement_id': False
     }
 
@@ -3642,9 +3731,17 @@ def renaming_item(ctx, source_path, new_name, apikey, cloudos_url,
 @click.option('--ssl-cert', help='Path to your SSL certificate file.')
 @click.option('--profile', default=None, help='Profile to use from the config file.')
 @click.pass_context
-def copy_item_cli(ctx, source_path, destination_path, apikey, cloudos_url,
-                  workspace_id, project_name, destination_project_name,
-                  disable_ssl_verification, ssl_cert, profile):
+def copy_item_cli(ctx,
+                  source_path,
+                  destination_path,
+                  apikey,
+                  cloudos_url,
+                  workspace_id,
+                  project_name,
+                  destination_project_name,
+                  disable_ssl_verification,
+                  ssl_cert,
+                  profile):
     """
     Copy a file or folder (S3 or virtual) from SOURCE_PATH to DESTINATION_PATH.
 
@@ -3653,6 +3750,7 @@ def copy_item_cli(ctx, source_path, destination_path, apikey, cloudos_url,
     DESTINATION_PATH [path]: the full path to the destination folder. It must be a 'Data' folder path.
      E.g.: Data/plots
     """
+    update_command_context_from_click(ctx)
     click.echo("Loading configuration profile...")
     config_manager = ConfigurationProfile()
     required_dict = {
@@ -3660,6 +3758,7 @@ def copy_item_cli(ctx, source_path, destination_path, apikey, cloudos_url,
         'workspace_id': True,
         'workflow_name': False,
         'project_name': True,
+        'session_id': False,
         'procurement_id': False
     }
     user_options = config_manager.load_profile_and_validate_data(
@@ -3746,9 +3845,9 @@ def copy_item_cli(ctx, source_path, destination_path, apikey, cloudos_url,
         click.echo(f"Copying {item_type.replace('_', ' ')} '{source_name}' to '{destination_path}'...")
         if destination_folder.get("folderType") is True and destination_folder.get("kind") in ("Data", "Cohorts", "AnalysesResults"):
             destination_kind = "Dataset"
-        elif destination_folder.get("folderType")=="S3Folder":
-            click.echo(f"[ERROR] Item '{source_name}' could not be copied to '{destination_path}' as the destination folder is not modifiable.",
-                   err=True)
+        elif destination_folder.get("folderType") == "S3Folder":
+            click.echo(f"[ERROR] Unable to copy item '{source_name}' to '{destination_path}'. The destination is an S3 folder, and only virtual folders can be selected as valid copy destinations.",
+                       err=True)
             sys.exit(1)
         else:
             destination_kind = "Folder"
@@ -3777,14 +3876,21 @@ def copy_item_cli(ctx, source_path, destination_path, apikey, cloudos_url,
 @click.option('--ssl-cert', help='Path to your SSL certificate file.')
 @click.option('--profile', default=None, help='Profile to use from the config file.')
 @click.pass_context
-def mkdir_item(ctx, new_folder_path, apikey, cloudos_url,
-               workspace_id, project_name,
-               disable_ssl_verification, ssl_cert, profile):
+def mkdir_item(ctx,
+               new_folder_path,
+               apikey,
+               cloudos_url,
+               workspace_id,
+               project_name,
+               disable_ssl_verification,
+               ssl_cert,
+               profile):
     """
     Create a virtual folder in a CloudOS project.
 
     NEW_FOLDER_PATH [path]: Full path to the new folder including its name. Must start with 'Data'.
     """
+    update_command_context_from_click(ctx)
     new_folder_path = new_folder_path.strip("/")
     if not new_folder_path.startswith("Data"):
         click.echo("[ERROR] NEW_FOLDER_PATH must start with 'Data'.", err=True)
@@ -3805,6 +3911,7 @@ def mkdir_item(ctx, new_folder_path, apikey, cloudos_url,
         'workspace_id': True,
         'workflow_name': False,
         'project_name': True,
+        'session_id': False,
         'procurement_id': False
     }
 
@@ -3896,16 +4003,25 @@ def mkdir_item(ctx, new_folder_path, apikey, cloudos_url,
 @click.option('--disable-ssl-verification', is_flag=True, help='Disable SSL certificate verification.')
 @click.option('--ssl-cert', help='Path to your SSL certificate file.')
 @click.option('--profile', default=None, help='Profile to use from the config file.')
+@click.option('--force', is_flag=True, help='Force delete files. Required when deleting user uploaded files. This may also delete the file from the cloud provider storage.')
 @click.pass_context
-def rm_item(ctx, target_path, apikey, cloudos_url,
-            workspace_id, project_name,
-            disable_ssl_verification, ssl_cert, profile):
+def rm_item(ctx,
+            target_path,
+            apikey,
+            cloudos_url,
+            workspace_id,
+            project_name,
+            disable_ssl_verification,
+            ssl_cert,
+            profile,
+            force):
     """
     Delete a file or folder in a CloudOS project.
 
     TARGET_PATH [path]: the full path to the file or folder to delete. Must start with 'Data'. \n
     E.g.: 'Data/folderA/file.txt' or 'Data/my_analysis/results/folderB'
     """
+    update_command_context_from_click(ctx)
     if not target_path.strip("/").startswith("Data/"):
         click.echo("[ERROR] TARGET_PATH must start with 'Data/', pointing to a file or folder.", err=True)
         sys.exit(1)
@@ -3916,6 +4032,7 @@ def rm_item(ctx, target_path, apikey, cloudos_url,
         'workspace_id': True,
         'workflow_name': False,
         'project_name': True,
+        'session_id': False,
         'procurement_id': False
     }
 
@@ -3971,21 +4088,32 @@ def rm_item(ctx, target_path, apikey, cloudos_url,
         click.echo(f"[ERROR] Item '{item_name}' not found in '{parent_path or '[project root]'}'", err=True)
         sys.exit(1)
 
-    item_id = found_item.get("_id",'')
+    item_id = found_item.get("_id", '')
     kind = "Folder" if "folderType" in found_item else "File"
-    if item_id=='':
-        click.echo(f"[ERROR] Item '{item_name}' could not be removed as the parent folder is not modifiable.",
+    if item_id == '':
+        click.echo(f"[ERROR] Item '{item_name}' could not be removed as the parent folder is an s3 folder and their content cannot be modified.",
                    err=True)
+        sys.exit(1)
+    # Check if the item is managed by Lifebit
+    is_managed_by_lifebit = found_item.get("isManagedByLifebit", False)
+    if is_managed_by_lifebit and not force:
+        click.echo("[ERROR] By removing this file, it will be permanently deleted. If you want to go forward, please use the --force flag.", err=True)
         sys.exit(1)
     click.echo(f"Removing {kind} '{item_name}' from '{parent_path or '[root]'}'...")
     try:
         response = client.delete_item(item_id=item_id, kind=kind)
         if response.ok:
-            click.secho(
-                f"[SUCCESS] {kind} '{item_name}' was removed from '{parent_path or '[root]'}'.",
-                fg="green", bold=True
-            )
-            click.secho("This item will still be available on your Cloud Provider.", fg="yellow")
+            if is_managed_by_lifebit:
+                click.secho(
+                    f"[SUCCESS] {kind} '{item_name}' was permanently deleted from '{parent_path or '[root]'}'.",
+                    fg="green", bold=True
+                )
+            else:
+                click.secho(
+                    f"[SUCCESS] {kind} '{item_name}' was removed from '{parent_path or '[root]'}'.",
+                    fg="green", bold=True
+                )
+                click.secho("This item will still be available on your Cloud Provider.", fg="yellow")
         else:
             click.echo(f"[ERROR] Removal failed: {response.status_code} - {response.text}", err=True)
             sys.exit(1)
@@ -4009,14 +4137,23 @@ def rm_item(ctx, target_path, apikey, cloudos_url,
 @click.option('--ssl-cert', help='Path to your SSL certificate file.')
 @click.option('--profile', help='Profile to use from the config file', default='default')
 @click.pass_context
-def link(ctx, path, apikey, cloudos_url, project_name, workspace_id, session_id, disable_ssl_verification, ssl_cert, profile):
+def link(ctx,
+         path,
+         apikey,
+         cloudos_url,
+         project_name,
+         workspace_id,
+         session_id,
+         disable_ssl_verification,
+         ssl_cert,
+         profile):
     """
     Link a folder (S3 or File Explorer) to an active interactive analysis.
 
-    PATH [path]: the full path to the S3 folder to link or relative to File Explorer. 
+    PATH [path]: the full path to the S3 folder to link or relative to File Explorer.
     E.g.: 's3://bucket-name/folder/subfolder', 'Data/Downloads' or 'Data'.
     """
-
+    update_command_context_from_click(ctx)
     profile = profile or ctx.default_map['datasets']['link']['profile']
 
     # Create a dictionary with required and non-required params
@@ -4025,6 +4162,7 @@ def link(ctx, path, apikey, cloudos_url, project_name, workspace_id, session_id,
         'workspace_id': True,
         'workflow_name': False,
         'project_name': False,
+        'session_id': True,
         'procurement_id': False
     }
     # determine if the user provided all required parameters
@@ -4068,7 +4206,6 @@ def link(ctx, path, apikey, cloudos_url, project_name, workspace_id, session_id,
     # Minimal folder validation and improved error messages
     is_s3 = path.startswith("s3://")
     is_folder = True
-    
     if is_s3:
         # S3 path validation - use heuristics to determine if it's likely a folder
         try:
@@ -4126,10 +4263,12 @@ def link(ctx, path, apikey, cloudos_url, project_name, workspace_id, session_id,
         if is_s3:
             click.echo("[ERROR] The S3 path appears to point to a file, not a folder. You can only link folders. Please link the parent folder instead.", err=True)
         else:
-            click.echo("[ERROR] Linking is only supported for folders, not individual files. Please link the parent folder instead.", err=True)
+            click.echo("[ERROR] Linking files or virtual folders is not supported. Link the S3 parent folder instead.", err=True)
         return
     elif is_folder is None and is_s3:
-        click.echo("[WARNING] Unable to verify whether the S3 path is a folder. Proceeding with linking; however, if the operation fails, please confirm that you are linking a folder rather than a file.", err=True)
+        click.echo("[WARNING] Unable to verify whether the S3 path is a folder. Proceeding with linking; " +
+                   "however, if the operation fails, please confirm that you are linking a folder rather than a file.",
+                   err=True)
 
     try:
         link_p.link_folder(path, session_id)
@@ -4170,7 +4309,7 @@ def list_images(ctx,
                 page,
                 limit):
     """List images associated with organisations of a given procurement."""
-
+    update_command_context_from_click(ctx)
     profile = profile or ctx.default_map['procurement']['images']['ls'].get('profile')
     config_manager = ConfigurationProfile()
 
@@ -4179,6 +4318,7 @@ def list_images(ctx,
         'workspace_id': False,
         'workflow_name': False,
         'project_name': False,
+        'session_id': False,
         'procurement_id': True
     }
 
@@ -4235,8 +4375,14 @@ def list_images(ctx,
               required=True)
 @click.option('--procurement-id', help='The specific CloudOS procurement id.', required=True)
 @click.option('--organisation-id', help='The Organisation Id where the change is going to be applied.', required=True)
-@click.option('--image-type', help='The CloudOS resource image type.', required=True, 
-            type=click.Choice(['RegularInteractiveSessions', 'SparkInteractiveSessions', 'RStudioInteractiveSessions', 'JupyterInteractiveSessions', 'JobDefault', 'NextflowBatchComputeEnvironment']))
+@click.option('--image-type', help='The CloudOS resource image type.', required=True,
+              type=click.Choice([
+                  'RegularInteractiveSessions',
+                  'SparkInteractiveSessions',
+                  'RStudioInteractiveSessions',
+                  'JupyterInteractiveSessions',
+                  'JobDefault',
+                  'NextflowBatchComputeEnvironment']))
 @click.option('--provider', help='The cloud provider. Only aws is supported.', required=True, type=click.Choice(['aws']), default='aws')
 @click.option('--region', help='The cloud region. Only aws regions are supported.', required=True)
 @click.option('--image-id', help='The new image id value.', required=True)
@@ -4263,7 +4409,7 @@ def set_organisation_image(ctx,
                            ssl_cert,
                            profile):
     """Set a new image id or name to image associated with an organisations of a given procurement."""
-
+    update_command_context_from_click(ctx)
     profile = profile or ctx.default_map['procurement']['images']['ls'].get('profile')
     config_manager = ConfigurationProfile()
 
@@ -4272,6 +4418,7 @@ def set_organisation_image(ctx,
         'workspace_id': False,
         'workflow_name': False,
         'project_name': False,
+        'session_id': False,
         'procurement_id': True
     }
 
@@ -4333,8 +4480,14 @@ def set_organisation_image(ctx,
               required=True)
 @click.option('--procurement-id', help='The specific CloudOS procurement id.', required=True)
 @click.option('--organisation-id', help='The Organisation Id where the change is going to be applied.', required=True)
-@click.option('--image-type', help='The CloudOS resource image type.', required=True, 
-            type=click.Choice(['RegularInteractiveSessions', 'SparkInteractiveSessions', 'RStudioInteractiveSessions', 'JupyterInteractiveSessions', 'JobDefault', 'NextflowBatchComputeEnvironment']))
+@click.option('--image-type', help='The CloudOS resource image type.', required=True,
+              type=click.Choice([
+                  'RegularInteractiveSessions',
+                  'SparkInteractiveSessions',
+                  'RStudioInteractiveSessions',
+                  'JupyterInteractiveSessions',
+                  'JobDefault',
+                  'NextflowBatchComputeEnvironment']))
 @click.option('--provider', help='The cloud provider. Only aws is supported.', required=True, type=click.Choice(['aws']), default='aws')
 @click.option('--region', help='The cloud region. Only aws regions are supported.', required=True)
 @click.option('--disable-ssl-verification',
@@ -4357,7 +4510,7 @@ def reset_organisation_image(ctx,
                              ssl_cert,
                              profile):
     """Reset image associated with an organisations of a given procurement to CloudOS defaults."""
-
+    update_command_context_from_click(ctx)
     profile = profile or ctx.default_map['procurement']['images']['set'].get('profile')
     config_manager = ConfigurationProfile()
 
@@ -4366,6 +4519,7 @@ def reset_organisation_image(ctx,
         'workspace_id': False,
         'workflow_name': False,
         'project_name': False,
+        'session_id': False,
         'procurement_id': True
     }
 
@@ -4414,14 +4568,18 @@ def reset_organisation_image(ctx,
 
 
 if __name__ == "__main__":
+    # Setup logging
+    debug_mode = '--debug' in sys.argv
+    setup_logging(debug_mode)
+    logger = logging.getLogger("CloudOS")
+    # Check if debug flag was passed (fallback for cases where Click doesn't handle it)
     try:
         run_cloudos_cli()
     except Exception as e:
-        # Check if debug flag was passed (fallback for cases where Click doesn't handle it)
-        debug_mode = '--debug' in sys.argv
-        
         if debug_mode:
+            logger.error(e, exc_info=True)
             traceback.print_exc()
         else:
+            logger.error(e)
             click.echo(click.style(f"Error: {e}", fg='red'), err=True)
         sys.exit(1)