+++ CHANGELOG.md
@@ -1,10 +1,31 @@
 ## lifebit-ai/cloudos-cli: changelog
 
-## v2.60.0 (2025-09-08)
+## v2.61.0 (2025-09-09)
 
 ### Feat
 
 - Adds job cost details in stdout, csv and json formats
+
+## v2.60.0 (2025-09-50)
+
+### Feat
+
+- Adds sphinx docs
+
+## v2.59.0 (2025-09-05)
+
+### Feat
+
+- Implements error logging
+
+## v2.58.1 (2025-09-04)
+
+### Patch
+
+- Change error message when copying files to s3 folder
+- Change error message when removing user-uploaded files
+- Raises error message if no session-id is provided to datasets link
+- Added some error messages for job workdir
 
 ## v2.58.0 (2025-09-01)